Implementing review changes and refactoring fixtures.

+++ cfme/tests/infrastructure/test_quota_tagging.py
@@ -22,9 +22,14 @@ pytestmark = [
 
 @pytest.fixture
 def admin_email(appliance):
+    """Required for user quota tagging services to work, as it's mandatory for it's functioning."""
     user = appliance.collections.users
     admin = user.instantiate(name='Administrator')
-    admin.update({'email': 'xyz@redhat.com'})
+    with update(admin):
+        admin.email = 'xyz@redhat.com'
+    yield
+    with update(admin):
+        admin.email = ''
 
 
 @pytest.fixture
@@ -55,14 +60,6 @@ def prov_data(provider, vm_name):
         }
 
 
-@pytest.fixture
-def custom_prov_data(request, prov_data, vm_name, template_name):
-    value = request.param
-    prov_data.update(value)
-    prov_data['catalog']['vm_name'] = vm_name
-    prov_data['catalog']['catalog_name'] = {'name': template_name}
-
-
 @pytest.fixture(scope='module')
 def test_domain(appliance):
     domain = appliance.collections.domains.create('test_{}'.format(fauxfactory.gen_alphanumeric()),
@@ -85,7 +82,8 @@ def catalog_item(appliance, provider, dialog, catalog, prov_data):
                                      dialog=dialog,
                                      prov_data=prov_data)
     yield catalog_item
-    catalog_item.delete()
+    if catalog_item.exists:
+        catalog_item.delete()
 
 
 @pytest.fixture(scope='module')
@@ -120,9 +118,7 @@ def set_entity_quota_source(max_quota_test_instance, entity):
 
 
 @pytest.fixture(params=[('groups', 'group', 'EvmGroup-super_administrator'),
-                        ('users', 'user', 'Administrator')],
-                ids=['group', 'user'],
-                scope='module')
+                        ('users', 'user', 'Administrator')], ids=['group', 'user'], scope='module')
 def entities(appliance, request, max_quota_test_instance):
     collection, entity, description = request.param
     set_entity_quota_source(max_quota_test_instance, entity)
@@ -151,9 +147,9 @@ def set_entity_quota_tag(request, entities, appliance):
     indirect=['set_entity_quota_tag'],
     ids=['max_memory', 'max_storage', 'max_cpu']
 )
-def test_quota_tagging_infra_via_lifecycle(appliance, provider, setup_provider,
-                                           set_entity_quota_tag, custom_prov_data, vm_name,
-                                           template_name, prov_data):
+def test_quota_tagging_infra_via_lifecycle(request, appliance, provider, setup_provider,
+                                           set_entity_quota_tag, custom_prov_data,
+                                           vm_name, template_name, prov_data):
     recursive_update(prov_data, custom_prov_data)
     do_vm_provisioning(appliance, template_name=template_name, provider=provider, vm_name=vm_name,
                        provisioning_data=prov_data, smtp_test=False, wait=False, request=None)
@@ -162,11 +158,12 @@ def test_quota_tagging_infra_via_lifecycle(appliance, provider, setup_provider,
     request_description = 'Provision from [{}] to [{}]'.format(template_name, vm_name)
     provision_request = appliance.collections.requests.instantiate(request_description)
     provision_request.wait_for_request(method='ui')
+    request.addfinalizer(provision_request.remove_request)
     assert provision_request.row.reason.text == "Quota Exceeded"
 
 
 @pytest.mark.rhv2
-@pytest.mark.parametrize('context', [ViaUI])
+@pytest.mark.parametrize('context', [ViaSSUI, ViaUI])
 @pytest.mark.parametrize(
     ['set_entity_quota_tag', 'custom_prov_data'],
     [
@@ -174,12 +171,13 @@ def test_quota_tagging_infra_via_lifecycle(appliance, provider, setup_provider,
         [('Quota - Max Storage *', '10GB'), {}],
         [('Quota - Max CPUs *', '1'), {'hardware': {'num_sockets': '8'}}]
     ],
-    indirect=['set_entity_quota_tag', 'custom_prov_data'],
+    indirect=['set_entity_quota_tag'],
     ids=['max_memory', 'max_storage', 'max_cpu']
 )
 def test_quota_tagging_infra_via_services(request, appliance, provider, setup_provider, admin_email,
-                                          context, set_entity_quota_tag, custom_prov_data, vm_name,
-                                          template_name, prov_data, catalog_item):
+                                          context, set_entity_quota_tag, custom_prov_data,
+                                          prov_data, catalog_item):
+    prov_data.update(custom_prov_data)
     with appliance.context.use(context):
         service_catalogs = ServiceCatalogs(appliance, catalog_item.catalog, catalog_item.name)
         if context is ViaSSUI:
@@ -189,8 +187,5 @@ def test_quota_tagging_infra_via_services(request, appliance, provider, setup_pr
     request_description = 'Provisioning Service [{0}] from [{0}]'.format(catalog_item.name)
     provision_request = appliance.collections.requests.instantiate(request_description)
     provision_request.wait_for_request(method='ui')
+    request.addfinalizer(provision_request.remove_request)
     assert provision_request.row.reason.text == "Quota Exceeded"
-
-    @request.addfinalizer
-    def delete():
-        provision_request.remove_request()
